q_libc: Avoid undefined behaviour with ctype functions.
The argument of these functions is of type int, but only a very
restricted subset of values are actually valid.  The argument must either
be the value of the macro EOF (which has a negative value), or must be a
non-negative value within the range representable as unsigned char.
Passing invalid values leads to undefined behavior.

Fixes a segmentation fault on NetBSD 11.